refactor(pools): reorganize error messages
