+++ src/utilities/erc20.ts
@@ -2,7 +2,7 @@ import { ethers } from 'ethers'
 
 import { miscErc20 } from '../artifacts'
 import { BaseContract } from '../base-classes/base-contract'
-import { ecpErrorMessage } from '../errors/error-messages'
+import { ecpErrorMessage } from '../errors'
 import { Abi, PrivateKey } from '../types/types'
 
 export class Erc20 extends BaseContract {
