+++ test/integration/self-provider-pools.ts
@@ -1,10 +1,7 @@
 import { isAddress, isError } from 'ethers'
 
 import { Erc20 } from '../../src'
-import {
-  cppErrorMessage,
-  poolCommonErrorMessage
-} from '../../src/errors/error-messages'
+import { cppErrorMessage, poolCommonErrorMessage } from '../../src/errors'
 import { Pools } from '../../src/pools/pools'
 import { SelfProvider } from '../../src/pools/self-provider'
 import { PoolInput, PoolStatusOption } from '../../src/types/pools'
